feat(goals): Implement year picker and support decimal rates for goal calculations
This commit enhances the goal creation and onboarding flows by introducing a dedicated year selection component and updating financial calculators to support decimal-based inflation and return rates for better precision.

**Key Changes:**

*   **Goal Calculations:**
    *   Updated `GoalCalculator`, `GoalRequest`, and related DTOs to use `Double` instead of `Int` for `inflationRate`, `returnRate`, and `postRetirementReturn`.
    *   Refactored `GoalRequest.getGoalInputs()` to return `Double` for rate-related values.

*   **Year Selection UI:**
    *   **`YearSelector.kt`:** Introduced a new `WheelPicker` and `YearPicker` component for a scrollable, native-feeling year selection experience.
    *   **Integration:** Replaced manual year input with the new `YearPicker` in both `AddGoalScreen` and `AddGoalOnboardingScreen`.

*   **Onboarding Enhancements:**
    *   **`OnBoardingRootScreen.kt`:** Added state management to handle the visibility of the new year picker.
    *   **`GoalScreenOnboardingViewModel.kt`:** Updated validation logic and data mapping to handle `Double` inputs for financial rates.
    *   **`AddGoalOnboardingScreen.kt`:** Automatically populates "Monthly Expense" based on previously calculated financial flow data and sets default retirement age from personal details.

*   **UI Refinements:**
    *   **Text Fields:** Updated `OnBoardingTextField` configurations for inflation and return rates to use `KeyboardType.Decimal` and implemented regex-based validation to ensure proper decimal formatting.
    *   **Layout Fixes:** Adjusted `CurrentAssetScreen` and `FinancialFlowScreen` to use `Column` with weighted `LazyColumn` for better scroll behavior and footer positioning.
    *   **Cleanup:** Temporarily commented out the `ExploreBundleGoalsSection` in `ProjectionImpactScreen`.

*   **Data Layer:**
    *   Updated `UserFinanceRepo` and various DTOs (`GoalRequestDto`, `ChildGoalBodyDto`) to align with the change from `Int` to `Double` for financial rates.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/data/remote/model/addgoals/ChildGoalBodyDto.kt
@@ -10,6 +10,6 @@ data class ChildGoalBodyDto(
     val current_saved_amount: Long,
     val goal_type_id: Int,
     val inflation_rate: Double,
-    val return_rate: Int,
+    val return_rate: Double,
     val years_left: Int
 )
